+++ examples/be-example-express-sqlite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # be-example-express-sqlite
 
+## 0.0.3
+
+### Patch Changes
+
+- Add langchain memory
+- Updated dependencies
+  - convostack@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes